+++ backend/src/utils/file/mod.rs
@@ -4,9 +4,11 @@ mod file_lock_manager;
 mod config_reader;
 mod env_resolving_reader;
 mod mapping_reader;
+mod template_reader;
 
 pub use self::file_utils::*;
 pub use self::file_lock_manager::*;
 pub use self::config_reader::*;
 pub use self::mapping_reader::*;
 pub use self::env_resolving_reader::*;
+pub use self::template_reader::*;