+++ frontend/src/app/components/config/config_view.rs
@@ -213,9 +213,11 @@ pub fn ConfigView() -> Html {
                     services.toastr.error(err.to_string());
                     return;
                 }
-                if let Err(err) =
-                    app_config.sources.prepare(false, app_config.config.get_hdhr_device_overview().as_ref())
-                {
+                if let Err(err) = app_config.sources.prepare(
+                    false,
+                    app_config.config.get_hdhr_device_overview().as_ref(),
+                    app_config.templates.as_ref().map(|defs| &defs.templates),
+                ) {
                     services.toastr.error(err.to_string());
                     return;
                 }
@@ -291,7 +293,7 @@ pub fn ConfigView() -> Html {
             }
 
             if let Some(sources) = modified_sources.as_mut() {
-                if let Err(err) = sources.prepare(false, None) {
+                if let Err(err) = sources.prepare(false, None, None) {
                     services.toastr.error(err.to_string());
                     return;
                 }