+++ frontend/src/app/components/config/main_config_view.rs
@@ -21,6 +21,7 @@ const LABEL_DISK_BASED_PROCESSING: &str = "LABEL.DISK_BASED_PROCESSING";
 const LABEL_WORKING_DIR: &str = "LABEL.WORKING_DIR";
 const LABEL_DEFAULT_USER_AGENT: &str = "LABEL.DEFAULT_USER_AGENT";
 const LABEL_MAPPING_PATH: &str = "LABEL.MAPPING_PATH";
+const LABEL_TEMPLATE_PATH: &str = "LABEL.TEMPLATE_PATH";
 const LABEL_BACKUP_DIR: &str = "LABEL.BACKUP_DIR";
 const LABEL_USER_CONFIG_DIR: &str = "LABEL.USER_CONFIG_DIR";
 const LABEL_SLEEP_TIMER_MINS: &str = "LABEL.SLEEP_TIMER_MINS";
@@ -41,6 +42,7 @@ generate_form_reducer!(
         WorkingDir => working_dir: String,
         DefaultUserAgent => default_user_agent: Option<String>,
         MappingPath => mapping_path: Option<String>,
+        TemplatePath => template_path: Option<String>,
         BackupDir => backup_dir: Option<String>,
         UserConfigDir => user_config_dir: Option<String>,
         SleepTimerMins => sleep_timer_mins: Option<u32>,
@@ -92,6 +94,7 @@ pub fn MainConfigView() -> Html {
                 { config_field!(form_state.form, translate.t(LABEL_WORKING_DIR), working_dir) }
                 { config_field_optional!(form_state.form, translate.t(LABEL_DEFAULT_USER_AGENT), default_user_agent) }
                 { config_field_optional!(form_state.form, translate.t(LABEL_MAPPING_PATH), mapping_path) }
+                { config_field_optional!(form_state.form, translate.t(LABEL_TEMPLATE_PATH), template_path) }
                 { config_field_optional!(form_state.form, translate.t(LABEL_BACKUP_DIR), backup_dir) }
                 { config_field_optional!(form_state.form, translate.t(LABEL_USER_CONFIG_DIR), user_config_dir) }
                 { config_field_optional!(form_state.form, translate.t(LABEL_SLEEP_TIMER_MINS), sleep_timer_mins) }
@@ -113,6 +116,7 @@ pub fn MainConfigView() -> Html {
                 { edit_field_text!(form_state, translate.t(LABEL_WORKING_DIR), working_dir, MainConfigFormAction::WorkingDir) }
                 { edit_field_text_option!(form_state, translate.t(LABEL_DEFAULT_USER_AGENT), default_user_agent, MainConfigFormAction::DefaultUserAgent) }
                 { edit_field_text_option!(form_state, translate.t(LABEL_MAPPING_PATH), mapping_path, MainConfigFormAction::MappingPath) }
+                { edit_field_text_option!(form_state, translate.t(LABEL_TEMPLATE_PATH), template_path, MainConfigFormAction::TemplatePath) }
                 { edit_field_text_option!(form_state, translate.t(LABEL_BACKUP_DIR), backup_dir, MainConfigFormAction::BackupDir) }
                 { edit_field_text_option!(form_state, translate.t(LABEL_USER_CONFIG_DIR), user_config_dir, MainConfigFormAction::UserConfigDir) }
                 { edit_field_number_option_u32!(form_state, translate.t(LABEL_SLEEP_TIMER_MINS), sleep_timer_mins, MainConfigFormAction::SleepTimerMins) }